requirements.txt
azure-identity==1.17.1
openai==1.35.4
pandas==2.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
streamlit==1.36.0
streamlit_chat==0.1.1
streamlit-code-editor==0.1.21
streamlit-echarts==0.4.0
azure-core==1.30.2
msal==1.29.0
msal-extensions==1.2.0
